requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
backports.zoneinfo==0.2.1
black==23.12.1
click==8.1.7
Django==4.2.8
djangorestframework==3.14.0
dnspython==2.4.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pymongo==4.6.1
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.9.0
tzlocal==5.2
